+++ node/resolvers/bindingInfo.ts
@@ -1,5 +1,6 @@
 export const bindingInfo = async (_: unknown, __: unknown, ctx: Context) => {
   const { clients } = ctx
+
   const { vbase } = clients
 
   try {