+++ CHANGELOG.md
@@ -11,6 +11,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [1.4.1] - 2021-12-09
 
+### Added
+- Block interface `BindingChallenge`
+- Block interface `CurrentBinding`
+- 2 new types of selectors, a `select` and a `list`
+- Flags to label bindings. It allows custom images or defaults to `defaultLocale` country code.
+- User location detection.
+
 ## [1.4.0] - 2021-10-05
 
 ### Changed